client: styling
