+++ estimatey-client/src/App.tsx
@@ -1,11 +1,15 @@
-import { BrowserRouter as Router } from "react-router-dom";
+import { Link, BrowserRouter as Router } from "react-router-dom";
 import AppRoutes from "./Routes";
+import appPaths from "./appPaths";
 
 const App = () => {
   return (
-    <Router>
-      <AppRoutes />
-    </Router>
+    <div className="p-4">
+      <Router>
+        <Link className="text-2xl underline" to={appPaths.home}>Home</Link>
+        <AppRoutes />
+      </Router>
+    </div>
   )
 }
